CORE-5120: Move `TestHttpClient` to `http-rpc-test-common`
