+++ libs/http-rpc/http-rpc-server-impl/src/integrationTest/kotlin/net/corda/httprpc/server/impl/HttpRpcServerRequestsTest.kt
@@ -8,8 +8,6 @@ import io.javalin.core.util.Header.WWW_AUTHENTICATE
 import net.corda.httprpc.server.apigen.test.TestJavaPrimitivesRPCopsImpl
 import net.corda.httprpc.server.config.models.HttpRpcSettings
 import net.corda.httprpc.server.impl.apigen.processing.openapi.schema.toExample
-import net.corda.httprpc.server.impl.utils.TestHttpClientUnirestImpl
-import net.corda.httprpc.server.impl.utils.WebRequest
 import net.corda.httprpc.server.impl.utils.multipartDir
 import net.corda.httprpc.test.*
 import net.corda.httprpc.tools.HttpVerb.DELETE
@@ -26,8 +24,10 @@ import org.junit.jupiter.api.Test
 import java.time.Instant
 import java.time.ZonedDateTime
 import kotlin.test.assertEquals
-import net.corda.httprpc.server.impl.utils.TestClientFileUpload
 import net.corda.httprpc.test.utls.ChecksumUtil
+import net.corda.httprpc.test.utls.TestClientFileUpload
+import net.corda.httprpc.test.utls.TestHttpClientUnirestImpl
+import net.corda.httprpc.test.utls.WebRequest
 
 class HttpRpcServerRequestsTest : HttpRpcServerTestBase() {
     companion object {
