+++ libs/http-rpc/http-rpc-test-common/build.gradle
@@ -14,11 +14,13 @@ dependencies {
     implementation "net.corda:corda-serialization"
     implementation "com.fasterxml.jackson.module:jackson-module-kotlin:$jacksonVersion"
     implementation "com.nimbusds:oauth2-oidc-sdk:$nimbusVersion"
+    implementation "com.konghq:unirest-java:$unirestVersion"
 
     implementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
 
     implementation project(":libs:http-rpc:http-rpc")
     implementation project(":libs:http-rpc:http-rpc-common")
+    implementation project(":libs:http-rpc:http-rpc-tools")
     implementation project(":libs:serialization:serialization-amqp")
     implementation project(":libs:lifecycle:lifecycle")
 }
